requirements.txt
django=5.2.1
djangorestframework=3.16.0
drf-spectacular=0.28.0
dotenv=0.9.9
psycopg2-binaryv=2.9.10